bum/requirements.txt

azure-functions
pandas==1.3.4
numpy==1.22.3
glob2==0.7
regex==2022.10.31
python-docx==0.8.11
openpyxl==3.0.9
flask==2.1.2
flask-restful==0.3.9
jsonify==0.5
requests==2.28.1
beautifulsoup4==4.11.1
streamlit
pytest==7.2.1
azure-storage-blob
spacy==3.4.0
waitress==2.1.2
Flask-Cors==3.0.10
pyyml==0.0.2
PyYAML==6.0
Flask-HTTPAuth==4.7.0
flask-swagger-ui==4.11.1
transformers==4.19.2
torch==1.13.1
scikit-learn==1.1.1
xlsxwriter==3.0.3
azure-ai-formrecognizer==3.2.0
mammoth==1.5.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.4.0/en_core_web_md-3.4.0.tar.gz/
